+++ migrations/m160320_093837__alter_table__cms_storage_file.php
@@ -8,10 +8,18 @@
 use yii\db\Schema;
 use yii\db\Migration;
 
-class m220319_093837__alter_table__cms_storage_file extends Migration
+class m160320_093837__alter_table__cms_storage_file extends Migration
 {
     public function safeUp()
     {
+        //m220319_093837__alter_table__cms_storage_file
+
+        if ($this->db->createCommand('SELECT * FROM migration WHERE version="m220319_093837__alter_table__cms_storage_file"')->queryOne())
+        {
+            $this->db->createCommand()->delete('migration', 'version = "m220319_093837__alter_table__cms_storage_file"')->execute();
+            return true;
+        }
+
         $this->dropForeignKey('storage_file_created_by', "{{%cms_storage_file}}");
         $this->dropForeignKey('storage_file_updated_by', "{{%cms_storage_file}}");